Resolved TypeScript error with `ChatMessageMention`
On TypeScript >= 3.9.5 the following error occurs when using the
`master` branch of the MS Graph typings:

```
node_modules/@microsoft/microsoft-graph-types/microsoft-graph.d.ts:13404:18 - error TS2430: Interface 'ChatMessageMention' incorrectly extends interface 'Entity'.
  Types of property 'id' are incompatible.
    Type 'number | undefined' is not assignable to type 'string | undefined'.
      Type 'number' is not assignable to type 'string | undefined'.

13404 export interface ChatMessageMention extends Entity {
```

This is because ChatMessageMention requires an `id` of `number`, but the
over-arching `Entity` class defines `id` as a `string`. This patch
corrects this issue removing the `Entity` association.